Moving equalizer into player screen and fixed combine playbacks in playlist setting not effective immediately
