+++ app/src/main/java/com/tachyonmusic/data/repository/PredefinedPlaylistsRepositoryImpl.kt
@@ -6,8 +6,7 @@ import com.tachyonmusic.domain.use_case.ObserveSettings
 import com.tachyonmusic.playback_layers.domain.PlaybackRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.plus
 
 class PredefinedPlaylistsRepositoryImpl(
@@ -15,10 +14,11 @@ class PredefinedPlaylistsRepositoryImpl(
     observeSettings: ObserveSettings,
     externalScope: CoroutineScope
 ) : PredefinedPlaylistsRepository {
-    override var songPlaylist: List<SinglePlayback> = emptyList()
-        private set
-    override var customizedSongPlaylist: List<SinglePlayback> = emptyList()
-        private set
+    private val _songPlaylist = MutableStateFlow<List<SinglePlayback>>(emptyList())
+    override val songPlaylist = _songPlaylist.asStateFlow()
+
+    private val _customizedSongPlaylist = MutableStateFlow<List<SinglePlayback>>(emptyList())
+    override val customizedSongPlaylist = _customizedSongPlaylist.asStateFlow()
 
     init {
         combine(
@@ -27,11 +27,11 @@ class PredefinedPlaylistsRepositoryImpl(
             observeSettings()
         ) { songs, customizedSongs, settings ->
             if (settings.combineDifferentPlaybackTypes) {
-                songPlaylist = songs + customizedSongs
-                customizedSongPlaylist = customizedSongs + songs
+                _songPlaylist.update { songs + customizedSongs }
+                _customizedSongPlaylist.update { customizedSongs + songs }
             } else {
-                songPlaylist = songs
-                customizedSongPlaylist = customizedSongs
+                _songPlaylist.update { songs }
+                _customizedSongPlaylist.update { customizedSongs }
             }
         }.launchIn(externalScope + Dispatchers.IO)
     }
