+++ app/src/main/java/com/tachyonmusic/domain/repository/PredefinedPlaylistsRepository.kt
@@ -1,8 +1,9 @@
 package com.tachyonmusic.domain.repository
 
 import com.tachyonmusic.core.domain.playback.SinglePlayback
+import kotlinx.coroutines.flow.StateFlow
 
 interface PredefinedPlaylistsRepository {
-    val songPlaylist: List<SinglePlayback>
-    val customizedSongPlaylist: List<SinglePlayback>
+    val songPlaylist: StateFlow<List<SinglePlayback>>
+    val customizedSongPlaylist: StateFlow<List<SinglePlayback>>
 }
